// src/models/npcs/actors/stats/professions/lower/types.ts
export type LowerStratum =
  | 'farmer'
  | 'herdsman'
  | 'fisherman'
  | 'miner'
  | 'lumberjack'
  | 'laborer'
  | 'servant'
  | 'criminal'
  | 'missionary'
  | 'sailor'
  | 'dock worker'
  | 'soldier (military)'
  | 'mercenary'
  | 'guard'
  | 'artist'
  | 'musician'

export type MercenarySpecs = 'brute' | 'defender' | 'marksman' | 'rogue' | 'sorcerer'